return random or specific worker for fileref fetch
By default MemPool will use a random worker on a node to fetch a `FileRef`.
Optionally `MemPool.enable_random_fref_serve[]` can be set to `false` to have it use a specific worker (the one with lowest worker id) for this role.
This is useful in cases where it desirable to have a specific process designated for IO tasks.

Also added tests for new methods.