Add DRef migration support
This commit allows DRefs to be migrated from their current worker to
another worker, and have future accesses to the DRef be seamlessly
redirected to the new owner. Multiple migrations of the same DRef may
occur, and a caching layer will ensure that redirects only add overhead
on the first access from each worker after a migration has occurred.